requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai==1.35.3
pgvector==0.2.5
python-multipart==0.0.9
pydantic==2.7.4
httpx==0.27.0
email_validator==2.1.1
jinja2==3.1.4
pytest==8.2.2
pytest-cov==5.0.0